redo custom lib and include path var assignment in .pro file
this is just to improve the windows .pro file and make it accept
lib and include paths etc on qmake command line.
This is needed for use with automated builds by sxc_winbuild.sh
against the same deps as .9 client.
Hopefully soon the 3 separate .pro files will be integrated to
a single file for win, nix, and osx, but for now this only affects
the windows version